Add --front and --url options.
Pluggable transport args are new in 0.2.5.1-alpha. If you try to use
them in 2.4, you get

Feb 03 20:19:43.261 [warn] Key digest for Bridge is wrong length.
Feb 03 20:19:43.261 [warn] Failed to parse/validate config: Bridge line did not parse. See logs for details.

The --front and --url options are the same as the front= and url= args.
The args are preferred over the options if both are present.