Add an --http-proxy option.
Causes HTTP requests to be made through the given proxy, which must be
given as a URL. For example,
	meek-client --http-proxy=http://127.0.0.1:8080/

Without this option, the default golang behavior of using a proxy iff
the HTTP_PROXY or http_proxy environment variable is set is used.

We found this useful in configuring meek-client to use another
circumvention system that works as an HTTP proxy.
https://lists.torproject.org/pipermail/tor-dev/2014-March/006356.html