Increase URL fetch timeout to 10 seconds.
The default timeout is 5 seconds.
	https://developers.google.com/appengine/docs/go/urlfetch/reference#Transport

Maybe this will reduce the number of failures I occasionally see. The
App Engine log when one of the failures happens is:
	[04/Mar/2014:09:31:48 -0800] "POST / HTTP/1.1" 500 77 - "Go 1.1 package http" "meek-reflect.appspot.com" ms=11 cpu_ms=0 cpm_usd=0.000089 app_engine_release=1.9.0
	RoundTrip: API error 2 (urlfetch: FETCH_ERROR)
The "ms=11" makes me think that it didn't run long enough to have
exceeded the deadline. But "API error 2" might have something to do with
the timeout, for example at the link
	https://groups.google.com/forum/#!topic/google-appengine-go/Qq3RbStW0gQ
When I run locally using the SDK with a short timeout, like 50
milliseconds, I get errors like this:
	2014/03/04 18:19:32 ERROR: RoundTrip: API error 2 (urlfetch: FETCH_ERROR): [Errno 101] Network is unreachable