Disable read/write deadlines when done reading or writing.
I was getting long hangs during bootstrapping on Windows 8 in a VM. It
seemed to happen most often right after receiving a payload of maximum
size, which mostly happens only during bootstrapping. Increasing the
timeout to 10 s made it go away. Even though we reset the timeout before
every read and write operation, my guess is there's some implicit read
or write that was reusing the timeout of the most recent operation. When
payloads were big, they took a little more time to process, and failed.