Read from the local tor in a goroutine.
Previously we were using a read timeout to control the polling interval.
We would try to read from the SOCKS port until the next poll was due; if
anything was read we would send it immediately, otherwise we would send
an empty request. In order to be able to read anything from the SOCKS
port, the timeout had to be set not too short: I had it at 100 ms for a
minimum. The downside of that was that when only downloading, and not
uploading, the minimum timeout would have to expire in between every
request, adding an extra 100 ms to every request.

So now we read from the SOCKS port in a goroutine, and have it write the
chunks it reads to a channel. The channel can be polled without waiting.