Also reset the poll interval to 0 when we had something to send.
There could be a delay, for example, when you first clicked a link after
letting the browser idle for a while. The local data from tor would be
read immediately, but it could be that nothing was availble to be read
from the remote peer immediately. There would be a wait up until the
next polling interval before any data was received.