Add some proxy support functions.
These are candidates to move to goptlib for proposal 232 support.

I assumed that you should be able to give a proxy host as a domain name,
but it turns out that proposal 232 doesn't actually say that
(https://trac.torproject.org/projects/tor/ticket/12125#comment:3). Some
of the tests use IP addresses and some use host names.