Comment #12146 caveat of Firefox helper.
People could get burned if they use an HTTP proxy with an unpatched
Firefox. ("Burned" in the same sense as if they made a plain Tor
connection--the censor knows they're using Tor and are trying to
circumvent.)

I'm not putting this fact into user-visible documentation because it
will affect few users, and be taken care of by a future Firefox release.
Those using the Browser Bundle are already covered. Only those
attempting to set up their own Firefox helper or BYOB setup will need to
know it.