Return an error string for request format errors.
Previously we just closed the socket. I was trying to use a plain http
URL through the helper, which isn't allowed by requestOK. (This is just
a safety feature because none of our deployment is planned to use plain
http.) It took me a little while to figure out what was going wrong. Now
you will get a line in the meek-client log:

error in handling request: helper returned error: request failed validation