Set network.proxy.socks_remote_dns=false inside the Firefox helper.
We need this setting to be false in order for DNS lookups to work
outside of a proxy, when the extension is run in Tor Browser
(https://trac.torproject.org/projects/tor/ticket/11183#comment:6).

By setting the pref inside the extension, we don't rely on it being set
in the user.js file, where it can instead have its default value of
true, which is safer if the browser somehow runs without the extension.
https://trac.torproject.org/projects/tor/ticket/12674#comment:2