Set Content-Type: application/octet-stream in meek-server.
Even thought it takes up space in the header, it looks like Google and
Amazon try to sniff the type and add the header anyway if it is missing.
(They set "text/plain; charset=utf-8" for zero-length bodies and
"application/octet-stream" for non-zero bodies). So we might as well add
it preemptively.

https://trac.torproject.org/projects/tor/ticket/12778